# out/src/PowerPuffBoys.SWKOM.Package.Services/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "name" : "SKS",
      "url" : "http://www.technikum-wien.at/"
    },
    "title" : "Parcel Logistics Service",
    "version" : "1.22.1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
...
  "paths" : {
...
    "/warehouse" : {
      "get" : {
        "operationId" : "exportWarehouses",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/warehouse"
                }
              }
            },
            "description" : "Successful response"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/error"
                }
              }
            },
            "description" : "The operation failed due to an error."
          },
          "404" : {
            "content" : { },
            "description" : "No hierarchy loaded yet."
          }
        },
        "summary" : "Exports the hierarchy of Warehouse and Truck objects.\n",
        "tags" : [ "warehouse-management" ]
      },
      "post" : {
        "operationId" : "importWarehouses",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/warehouse"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "Successfully loaded."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/error"
                }
              }
            },
            "description" : "The operation failed due to an error."
          }
        },
        "summary" : "Imports a hierarchy of Warehouse and Truck objects.\n",
        "tags" : [ "warehouse-management" ]
      }
    },
    "/warehouse/{code}" : {
      "get" : {
        "operationId" : "getWarehouse",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "code",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/hop"
                }
              }
            },
            "description" : "Successful response"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/error"
                }
              }
            },
            "description" : "The operation failed due to an error."
          },
          "404" : {
            "description" : "No hop with the specified id could be found."
          }
        },
        "summary" : "Get a certain warehouse or truck by code",
        "tags" : [ "warehouse-management" ]
      }
    }
  },
  "components" : {
...
    "schemas" : {
...
      "hop" : {
        "discriminator" : {
          "propertyName" : "hopType"
        },
        "example" : {
          "code" : "code",
          "locationName" : "locationName",
          "processingDelayMins" : 0,
          "hopType" : "hopType",
          "description" : "description",
          "locationCoordinates" : {
            "lon" : 1.4658129805029452,
            "lat" : 6.027456183070403
          }
        },
        "properties" : {
          "hopType" : {
            "type" : "string"
          },
          "code" : {
            "description" : "Unique CODE of the hop.",
            "pattern" : "^[A-Z]{4}\\d{1,4}$",
            "type" : "string"
          },
          "description" : {
            "description" : "Description of the hop.",
            "type" : "string"
          },
          "processingDelayMins" : {
            "description" : "Delay processing takes on the hop.",
            "type" : "integer"
          },
          "locationName" : {
            "description" : "Name of the location (village, city, ..) of the hop.",
            "type" : "string"
          },
          "locationCoordinates" : {
            "$ref" : "#/components/schemas/geoCoordinate"
          }
        },
        "required" : [ "code", "description", "hopType", "locationCoordinates", "locationName", "processingDelayMins" ],
        "type" : "object"
      },
      "warehouse" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/hop"
        }, {
          "$ref" : "#/components/schemas/warehouse_allOf"
        } ],
        "required" : [ "code", "description", "hopType", "level", "locationCoordinates", "locationName", "nextHops", "processingDelayMins" ],
        "type" : "object"
      },
...
      "error" : {
        "properties" : {
          "errorMessage" : {
            "description" : "The error message.",
            "type" : "string"
          }
        },
        "required" : [ "errorMessage" ],
        "type" : "object"
      },
...
      "geoCoordinate" : {
        "example" : {
          "lon" : 1.4658129805029452,
          "lat" : 6.027456183070403
        },
        "properties" : {
          "lat" : {
            "description" : "Latitude of the coordinate.",
            "format" : "double",
            "type" : "number"
          },
          "lon" : {
            "description" : "Longitude of the coordinate.",
            "format" : "double",
            "type" : "number"
          }
        },
        "required" : [ "lat", "lon" ],
        "type" : "object"
      },
      "warehouse_allOf_nextHops" : {
        "properties" : {
          "traveltimeMins" : {
            "type" : "integer"
          },
          "hop" : {
            "$ref" : "#/components/schemas/hop"
          }
        },
        "required" : [ "hop", "traveltimeMins" ],
        "type" : "object",
        "example" : null
      },
      "warehouse_allOf" : {
        "properties" : {
          "level" : {
            "format" : "int32",
            "type" : "integer"
          },
          "nextHops" : {
            "description" : "Next hops after this warehouse (warehouses or trucks).",
            "items" : {
              "$ref" : "#/components/schemas/warehouse_allOf_nextHops"
            },
            "type" : "array"
          }
        },
        "example" : null
      },
...
    }
  }
}